Removing extra files
